+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `add` **Introduce `Extension::common_prepend_methods` to prepend methods to an instance and class.**
+
+    *Related links:*
+    - [Pull Request #419][pr-419]
+
   * `fix` **Ensure that defined state maintains the original definition order.**
 
     *Related links:*
@@ -218,6 +223,7 @@
     *Related links:*
     - [Pull Request #364][pr-364]
 
+[pr-419]: https://github.com/pakyow/pakyow/pull/419
 [pr-414]: https://github.com/pakyow/pakyow/pull/414
 [pr-408]: https://github.com/pakyow/pakyow/pull/408
 [pr-403]: https://github.com/pakyow/pakyow/pull/403